LayoutTests/inspector/console-dirxml.html

<html>
<head>
<script src="inspector-test.js"></script>
<script src="console-tests.js"></script>
<script>

function doit()
{
    console.dirxml(document);

    dumpConsoleMessages();
}

</script>
</head>

<body onload="onload()">
<p>
Tests that console logging dumps proper messages.
</p>

</body>
</html>
